Corrected title in the man of speedcomp.
Man for workingdir now contains a windows-specific part with a link to
setWindowTitle. Removed the example.

Package man file now has a correct WWW address.

Updated man page of sourceall. More backslahes in the default value of pattern
argument.

NAMESPACE was updated accordingly to the removal of some of the functions.


git-svn-id: file:///home/mbojan/svn/icm/R/mbtools/trunk@1100 c6088350-5b22-44b5-8b47-a67d22253714